Changed MyCommand to RelayCommand. Fully implemented RelayCommand class.
Added Command to ViewModel for Translate Button to bind to.
Added TranslateMorse object to viewmodel for calling methods from said class.
Added method to ViewModel for handling ToMorse Translation. (It will first check if the string is NullOrEmpty before calling ToMorse method.)

Implemented binding from View objects to ViewModel.

And tested it all which seems to work so far.